migrate-sfm-rules-main/requirements.txt
anyio==3.7.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
faker==24.11.0
faker_airtravel==0.4
Flask==3.0.2
googletrans==4.0.0rc1
h11==0.9.0
httpcore==0.9.1
httpx==0.23.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
kafka-python==2.0.2
MarkupSafe==2.1.5
numpy==1.26.4
openpyxl==3.1.2
packaging==23.2
pandas~=2.2.2
pillow==10.3.0
pluggy==1.4.0
pre-commit == 3.7.0
pyarrow==15.0.0
pytest==8.0.1
pytest-html==4.1.1
pytest-html-reporter==0.2.9
pytest-metadata==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlalchemy
tabulate==0.9.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1

xmltodict~=0.13.0
